# backend-1/src/open-api/swagger.yaml
openapi: 3.0.0
info:
  title: IoT API
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/diagnostics:
    get:
      summary: Get device diagnostics
      description: >-
        Returns device diagnostics including port statistics. Data is cached in
        Redis for performance.
      tags:
        - Diagnostics
      responses:
        '200':
          description: Diagnostics data with port statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  PortStatistics:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        description:
                          type: string
                        channel:
                          type: string
                        severity:
                          type: string
                          enum:
                            - Info
                            - Warning
                            - Maintenance requested
                            - Fault
                            - Critical
                        type:
                          type: string
                          enum:
                            - Occured
                            - Cleared
                            - Active
                            - Inactive
                        time:
                          type: string
        '500':
          description: Internal server error
    put:
      summary: Update diagnostics data
      description: >-
        Updates device diagnostics and invalidates Redis cache. Requires
        authentication.
      tags:
        - Diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                PortStatistics:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Updated diagnostics data
        '404':
          description: Diagnostics not found
        '500':
          description: Failed to update diagnostics
    post:
      summary: Add new diagnostic entry
      description: Adds a new port statistic entry to diagnostics. Requires authentication.
      tags:
        - Diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                description:
                  type: string
                channel:
                  type: string
                severity:
                  type: string
                type:
                  type: string
                time:
                  type: string
      responses:
        '201':
          description: Diagnostic entry added successfully
        '400':
          description: Invalid request data
        '500':
          description: Failed to add diagnostic entry
    delete:
      summary: Clear all diagnostics
      description: Clears all port statistics. Requires authentication.
      tags:
        - Diagnostics
      responses:
        '200':
          description: Diagnostics cleared successfully
        '404':
          description: Diagnostics not found
        '500':
          description: Failed to clear diagnostics
  /api/device-info:
    get:
      summary: Get device information
      description: >-
        Returns basic information about the device. Data is cached in Redis for
        performance.
      tags:
        - Device
      responses:
        '200':
          description: JSON object containing device information
          content:
            application/json:
              schema:
                type: object
                example:
                  name: Device 1
                  location: Berlin
                  type: MVK Pro
                  model: A100
                  status: online
        '500':
          description: Internal server error
    put:
      summary: Update device information
      description: Updates the device information and invalidates Redis cache.
      tags:
        - Device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                name: New Device Name
                location: Hanoi
                model: B200
                status: offline
      responses:
        '200':
          description: Updated device information
          content:
            application/json:
              schema:
                type: object
                example:
                  name: New Device Name
                  location: Hanoi
                  model: B200
                  status: offline
        '404':
          description: Device not found
        '500':
          description: Failed to update device
...